rts: rename interop-activity-logging to interop
